+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.0
+
+- Account registration goal added
+- Fix checkout goal JS step error
+- Allow to send multiple events on the same page
+
 ## 1.0.1
 
 - Support for self-hosted plausible instance ([@atmaniak](https://github.com/atmaniak))